% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCellig.R
\name{run_HCellig}
\alias{run_HCellig}
\title{HCellig is a hierarchical cell identity-based method
It considers three hierarchical layers: cell type, subtype and phenotype
It quantifies different levels of gene expression to accurately identify cell identity genes}
\usage{
run_HCellig(
  data,
  metadata,
  org = "human",
  dir,
  file.name,
  precBack = NA,
  discretize = TRUE,
  sig.frames = TRUE,
  ncores = detectCores() - 2,
  fixseed = 1234
)
}
\arguments{
\item{data}{single cell expression matrix}

\item{metadata}{corresponding metadata (two columns: cell names, cell type)}

\item{org}{organism (human or mouse)}

\item{dir}{directory used to save the results}

\item{file.name}{prefix for output files and folder}

\item{precBack}{precompiled background path to ScRbRef}

\item{discretize}{gene quantification in three levels of cell types}

\item{sig.frames}{construction of significant genes data frame}

\item{ncores}{number of cores to use to run the code}

\item{fixseed}{seed used to discretize cell types (reproducbility)}
}
\value{
NULL, all results are accessible in the dir specified
}
\description{
HCellig is a hierarchical cell identity-based method
It considers three hierarchical layers: cell type, subtype and phenotype
It quantifies different levels of gene expression to accurately identify cell identity genes
}
\author{
Céline Barlier, Kartikeya Singh, Sascha Jung
}
